OneToMany: fixed bug; remove nesmaze child kdyz ten dostane (pred persistovanim) nahradni parent
Pri OneToMany::remove muze smazat child entitu, kdyz bez vazby nemuze existovat.
To sice neni nejidealnejsi chovani, ale je to tak historicky dano.
Oprava bugu spociva v tom ze nebylo mozne nastavit te child entite jineho parent,
s kterym by mohla existovat. OneToMany ji stejne smazala.

+++ Orm/Relationships/OneToMany.php
@@ -250,9 +250,24 @@ class OneToMany extends BaseToMany implements IRelationship
 	{
 		$repository = $this->getChildRepository();
 
-		foreach ($this->del as $entity)
+		if ($this->del)
 		{
-			$repository->remove($entity);
+			$param = $this->getMetaData()->getChildParam();
+			$parent = $this->getParent();
+			foreach ($this->del as $entity)
+			{
+				if (!isset($entity->{$param}) OR $entity->{$param} === $parent)
+				{
+					$repository->remove($entity);
+				}
+				else
+				{
+					if ($all OR !isset($entity->id))
+					{
+						$repository->persist($entity, $all);
+					}
+				}
+			}
 		}
